requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
feedgenerator==1.9
google-api-core==1.14.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
idna==2.8
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3